requirements.txt
python-telegram-bot[job-queue]>=21.0,<22
gspread>=6.0.0,<7
google-auth>=2.0.0,<3